requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
sentence-transformers
faiss-cpu
pdfplumber
tqdm
numpy
requests
